Changed "kira" to "kit."
Apparently "Kira" is not just some Star Trek name, it's also
known to people that watch Japanese TV shows that are made
for children.  "Oh no, it's SEINEN, for the older audience!"
-- says the obese mid-20's male virgin whose greatest life
accomplishment was not being a brony.  (It's fine to have
weird tastes, but not when you have them without shame.)