s1: Typed slices instead of foo_count, foo_limit tuplets.
Originally, I used the SLICE_* macros in this throwaway-code personal
project in order to see how bad it could really be.  But now I want to
look for a job.  I'm worried that people would see that code and judge
me negatively for it.  So slices are getting migrated to something a
bit more sensible.  There are still ungrowable (ptr, count) pairs
being stored in separate values.  But the grotesque (ptr, count,
limit) triads are gone.